+++ click_extra/tests/test_colorize.py
@@ -1,0 +1,72 @@
+# Copyright Kevin Deldycke <kevin@deldycke.com> and contributors.
+# All Rights Reserved.
+#
+# This program is Free Software; you can redistribute it and/or
+# modify it under the terms of the GNU General Public License
+# as published by the Free Software Foundation; either version 2
+# of the License, or (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program; if not, write to the Free Software
+# Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
+
+from boltons.strutils import strip_ansi
+
+from ..colorize import HelpExtraFormatter, theme
+
+
+def test_options_highlight():
+    formatter = HelpExtraFormatter()
+    formatter.write("applies filtering by --manager and --exclude options")
+
+    formatter.long_options = {"--manager", "--exclude"}
+
+    output = formatter.getvalue()
+    assert theme.option("--manager") in output
+    assert theme.option("--exclude") in output
+
+
+def test_choices_highlight():
+    formatter = HelpExtraFormatter()
+    formatter.write(
+        """
+        -e, --apt [apm|apt|apt-mint|brew]
+                        Exclude a package manager.
+                        Repeat to exclude multiple
+                        managers.
+        """
+    )
+
+    formatter.choices = {"apm", "apt", "apt-mint", "brew"}
+
+    output = formatter.getvalue()
+    assert theme.choice("apm") in output
+    assert theme.choice("apt") in output
+    assert theme.choice("apt-mint") in output
+    assert theme.choice("brew") in output
+
+
+def test_metavars_highlight():
+    formatter = HelpExtraFormatter()
+    formatter.write("-v, --verbosity LEVEL   Either CRITICAL, ERROR or DEBUG.")
+
+    formatter.metavars = {"LEVEL"}
+
+    output = formatter.getvalue()
+    assert theme.metavar("LEVEL") in output
+
+
+def test_only_full_word_highlight():
+    formatter = HelpExtraFormatter()
+    formatter.write("package snapshot")
+
+    formatter.choices.add("snap")
+
+    output = formatter.getvalue()
+    # Make sure no highlighting occured
+    assert strip_ansi(output) == output